backend/requirements.txt
django==3.1
djangorestframework==3.11.1
djangorestframework_simplejwt==4.4.0
psycopg2-binary==2.8.5
python-dotenv==0.14.0
gunicorn==20.0.4
imagekitio
django-unixtimestampfield